requirements.txt
alembic
python-dotenv
psycopg2
fastapi
bcrypt
uvicorn
sqlalchemy
passlib
pyjwt
pydantic
pydantic[email]
